Convert ProfileBarChart.js to ES6 modules
